Moved tcp assertions, changed timezones, added static time and description
