+++ src/main/java/com/axibase/tsd/api/transport/tcp/TCPSender.java
@@ -72,27 +72,4 @@ public class TCPSender {
         send(commands, false);
     }
 
-    public static void assertBadTcpResponse(String response) {
-        if(response.equals("ok")) {
-            throw new AssertionError();
-        }
-    }
-
-    public static void assertBadTcpResponse(String message,String response) {
-        if(response.equals("ok")) {
-            throw new AssertionError(message);
-        }
-    }
-
-    public static void assertGoodTcpResponse(String response) {
-        if(!response.equals("ok")) {
-            throw new AssertionError();
-        }
-    }
-
-    public static void assertGoodTcpResponse(String message,String response) {
-        if(!response.equals("ok")) {
-            throw new AssertionError(message);
-        }
-    }
 }
